# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Canvica Cleaning Services CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /public/inquiry:
    post:
      operationId: submitInquiry
      tags: [public]
      summary: Submit a booking inquiry (creates a lead)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InquiryInput"
      responses:
        "201":
          description: Inquiry submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /public/calculate-quote:
    post:
      operationId: calculateQuotePublic
      tags: [public]
      summary: Calculate a quote from property details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteCalculatorInput"
      responses:
        "200":
          description: Quote estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteEstimate"

  # ─── DASHBOARD ──────────────────────────────────────────────────────────────
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads with optional filters
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [new, contacted, quoted, won, lost]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a lead manually
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /leads/{id}/convert:
    post:
      operationId: convertLead
      tags: [leads]
      summary: Convert a lead to a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Customer created from lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
...
components:
  schemas:
...
    Lead:
      type: object
      required: [id, name, email, phone, status, serviceType, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: ["string", "null"]
        serviceType:
          type: string
          enum: [residential, commercial]
        propertyType:
          type: ["string", "null"]
        bedrooms:
          type: ["integer", "null"]
        bathrooms:
          type: ["integer", "null"]
        squareFootage:
          type: ["integer", "null"]
        cleaningType:
          type: ["string", "null"]
        frequency:
          type: ["string", "null"]
        preferredDate:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        status:
          type: string
          enum: [new, contacted, quoted, won, lost]
        referralSource:
          type: ["string", "null"]
        estimatedValue:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    LeadInput:
      type: object
      required: [name, email, phone, serviceType]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        serviceType:
          type: string
          enum: [residential, commercial]
        propertyType:
          type: string
        bedrooms:
          type: number
        bathrooms:
          type: number
        squareFootage:
          type: number
        cleaningType:
          type: string
        frequency:
          type: string
        preferredDate:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [new, contacted, quoted, won, lost]
        referralSource:
          type: string
        estimatedValue:
          type: number
        notes:
          type: string
    LeadUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        serviceType:
          type: string
          enum: [residential, commercial]
        propertyType:
          type: string
        bedrooms:
          type: number
        bathrooms:
          type: number
        squareFootage:
          type: number
        cleaningType:
          type: string
        frequency:
          type: string
        preferredDate:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [new, contacted, quoted, won, lost]
        referralSource:
          type: string
        estimatedValue:
          type: number
        notes:
          type: string
    InquiryInput:
      type: object
      required: [name, email, phone, serviceType]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        serviceType:
          type: string
          enum: [residential, commercial]
        propertyType:
          type: string
        bedrooms:
          type: number
        bathrooms:
          type: number
        squareFootage:
          type: number
        cleaningType:
          type: string
        frequency:
          type: string
        preferredDate:
          type: string
        message:
          type: string
        referralSource:
          type: string
...
    Customer:
      type: object
      required: [id, name, email, phone, status, serviceType, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: ["string", "null"]
        gpsLat:
          type: ["number", "null"]
        gpsLng:
          type: ["number", "null"]
        serviceType:
          type: string
          enum: [residential, commercial]
        propertyType:
          type: ["string", "null"]
        bedrooms:
          type: ["integer", "null"]
        bathrooms:
          type: ["integer", "null"]
        squareFootage:
          type: ["integer", "null"]
        entryInstructions:
          type: ["string", "null"]
        alarmCode:
          type: ["string", "null"]
        gateCode:
          type: ["string", "null"]
        pets:
          type: ["string", "null"]
        preferredProducts:
          type: ["string", "null"]
        cleaningPreferences:
          type: ["string", "null"]
        specialInstructions:
          type: ["string", "null"]
        recurringSchedule:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, prospect, vip]
        referralSource:
          type: ["string", "null"]
        lifetimeValue:
          type: number
        isRecurring:
          type: boolean
        satisfactionScore:
          type: ["number", "null"]
        reviewScore:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    QuoteLineItem:
      type: object
      required: [description, quantity, unitPrice, total]
      properties:
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        total:
          type: number
...
    QuoteCalculatorInput:
      type: object
      required: [serviceType, cleaningType]
      properties:
        serviceType:
          type: string
          enum: [residential, commercial]
        cleaningType:
          type: string
          enum: [standard, deep, move_in_out, post_construction, recurring]
        bedrooms:
          type: number
        bathrooms:
          type: number
        squareFootage:
          type: number
        extras:
          type: array
          items:
            type: string
        frequency:
          type: string
          enum: [one_time, weekly, biweekly, monthly]
    QuoteEstimate:
      type: object
      required: [basePrice, extras, subtotal, tax, total, lineItems, notes]
      properties:
        basePrice:
          type: number
        extras:
          type: array
          items:
            type: object
            required: [name, price]
            properties:
              name:
                type: string
              price:
                type: number
        subtotal:
          type: number
        taxRate:
          type: number
        tax:
          type: number
        total:
          type: number
        discountPercent:
          type: number
        discountAmount:
          type: number
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/QuoteLineItem"
        notes:
          type: string

    # ─── INVOICE SCHEMAS ──────────────────────────────────────────────────────
...